Code refactoring.
